requirements.txt
# Azure dependencies
azure-identity>=1.19.0
azure-cosmos>=4.9.0
azure-search-documents>=11.6.0
openai>=1.59.0

# Azure agent-framework
agent-framework>=1.0.0b251001  # Microsoft Agent Framework SDK
agent-framework-azure-ai>=1.0.0b251001 

# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# Utilities
python-dotenv>=1.0.0
httpx>=0.27.0
tenacity>=9.0.0

# Data processing
numpy>=1.26.0
pandas>=2.2.0

# Testing
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
